copy resources/figs if it exists only

+++ R/render.R
@@ -5,7 +5,8 @@ system('mkdir -p render')
 system('mkdir render/courses_tools') 
 system('mkdir render/resources') 
 system( glue::glue( 'cp -rf courses_tools/resources render/courses_tools/.'))
-system( glue::glue( 'cp -rf resources/figs render/resources/.'))
+if(file.exists('resources/figs'))
+  system( glue::glue( 'cp -rf resources/figs render/resources/.'))
 
 source(file.path('courses_tools','R', 'load_packages.R'))
